+++ src/init.c
@@ -125,7 +125,6 @@ R_CallMethodDef callMethods[] = {
 {"Cinrange", (DL_FUNC) &inrange, -1},
 {"Cbetween", (DL_FUNC) &between, -1},
 {"ChasOpenMP", (DL_FUNC) &hasOpenMP, -1},
-{"CbeforeR340", (DL_FUNC) &beforeR340, -1},
 {"CuniqueNlogical", (DL_FUNC) &uniqueNlogical, -1},
 {"CfrollfunR", (DL_FUNC) &frollfunR, -1},
 {"CdllVersion", (DL_FUNC) &dllVersion, -1},
@@ -351,15 +350,6 @@ SEXP hasOpenMP(void) {
 
 }
 
-SEXP beforeR340(void) {
-  // used in onAttach.R for message about fread memory leak fix needing R 3.4.0
-  // at C level to catch if user upgrades R but does not reinstall data.table
-  #if R_VERSION < R_Version(3,4,0)
-  return ScalarLogical(true);
-  #else
-  return ScalarLogical(false);
-  #endif
-}
 // # nocov end
 
 extern int *_Last_updated;  // assign.c